requirements.txt
langchain
python-dotenv
langchain-community
langchain_groq
langchain_core
streamlit 
